requirements.txt
fastapi
uvicorn
gunicorn
sentence-transformers
openai
requests
python-dotenv
redis
